requirements.txt







## USE THE install.sh SCRIPT




absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.10.0
grpcio==1.67.0
huggingface-hub==0.26.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lightning==2.4.0
lightning-utilities==0.11.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.1
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
protobuf==5.28.2
psutil==6.1.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
timm==1.0.11
torch~=2.4.0
torchvision~=0.19.0
torchmetrics==1.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
umap-learn~=0.5.7
urllib3==2.2.3
yarl==1.15.5

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
rootutils~=1.0.7
pre-commit      # hooks for applying linters on commit
rich~=13.9.3
pytest~=8.3.3

# -------- tensor manipulation
einops~=0.8.0


opencv-python~=4.10.0.84
h5py~=3.12.1
omegaconf~=2.3.0
torch-geometric~=2.6.1
aim~=3.25.1
click~=8.1.7
setuptools~=75.1.0
pydantic~=2.9.2
prompt_toolkit
x_transformers~=1.42.18

hilbertcurve~=2.0.5
prompt-toolkit~=3.0.48
readline~=8.2
scikit-image~=0.24.0
pacmap~=0.7.6
einx~=0.3.0
loguru~=0.7.2
plotly~=5.24.1
dash~=2.18.2
dash-bootstrap-components~=1.6.0